Moved ChatRoom app files
