+++ src/krow/kröw/core/managers/ProgramSettings.java
@@ -12,8 +12,8 @@ import java.nio.file.Paths;
 
 import kröw.core.Kröw;
 import kröw.gui.ApplicationManager;
-import zeale.guis.ChatRoom;
 import zeale.guis.Statistics;
+import zeale.guis.chatroom.ChatRoom;
 
 public class ProgramSettings implements Serializable {
